Moved Cache trait to mod.rs
