+++ src/cache/cache_op_executors.rs
@@ -1,6 +1,6 @@
-use crate::cache::cache::Cache;
 use crate::cache::cache_operations::{CacheResponse, InvalidationPattern, RequestCached};
 use crate::cache::inner_cache::CachedWithCode;
+use crate::cache::Cache;
 use crate::utils::errors::{ApiError, ApiResult};
 use rocket::response::content;
 use serde::Serialize;
