+++ src/cache/mod.rs
@@ -5,3 +5,18 @@ mod inner_cache;
 
 #[cfg(test)]
 mod tests;
+
+use mockall::automock;
+
+#[automock]
+pub trait Cache {
+    fn fetch(&self, id: &str) -> Option<String>;
+    fn create(&self, id: &str, dest: &str, timeout: usize);
+    fn insert_in_hash(&self, hash: &str, id: &str, dest: &str);
+    fn get_from_hash(&self, hash: &str, id: &str) -> Option<String>;
+    fn has_key(&self, id: &str) -> bool;
+    fn expire_entity(&self, id: &str, timeout: usize);
+    fn invalidate_pattern(&self, pattern: &str);
+    fn invalidate(&self, id: &str);
+    fn info(&self) -> Option<String>;
+}
