+++ src/routes/about.rs
@@ -1,5 +1,5 @@
-use crate::cache::cache::Cache;
 use crate::cache::cache_operations::CacheResponse;
+use crate::cache::Cache;
 use crate::config::base_transaction_service_url;
 use crate::config::{about_cache_duration, webhook_token};
 use crate::services::about;
