+++ src/services/hooks.rs
@@ -1,5 +1,5 @@
-use crate::cache::cache::Cache;
 use crate::cache::cache_operations::{Invalidate, InvalidationPattern};
+use crate::cache::Cache;
 use crate::models::backend::webhooks::{Payload, PayloadDetails};
 use crate::utils::errors::ApiResult;
